edam

+++ src/database/schemas/Uploaders.cs
@@ -2,7 +2,7 @@ namespace net.niceygy.eddatacollector.database.schemas
 {
     public class Uploaders
     {
-        public required int timestamp { get; set; }
+        public required long timestamp { get; set; }
         public required int uploaders { get; set; }
     }
 }
